DAVCPP-33: refactored `Value` constructors by removing explicit `ValueType` parameters; added `tryPopStackValue` to `DavScriptVirtualMachine`; introduced `@param` token type and updated corresponding logic; enhanced `print` function to include optional type information; fixed minor lexer and parser issues; added tests for new functionality

+++ DaVinCpp2_0/src/davscript/parser/ast/ValueNode.cpp
@@ -61,7 +61,7 @@ namespace davincpp::davscript
             }
             case FLOATTYPE:
             {
-                valueBytes = ByteCastHelper::nativeToBytes(std::stol(actualValue));
+                valueBytes = ByteCastHelper::nativeToBytes(std::stod(actualValue));
                 break;
             }
             case STRINGTYPE:
